requirements.txt
google-generativeai
python-dotenv
streamlit
pillow
chromadb
pyPDF2
langchain
faiss-cpu
langchain_google_genai
